refactor(canon): split virtual_ts/generator into cohesive modules
