+++ crates/vize_canon/src/virtual_ts/generator/mod.rs
@@ -3,8 +3,23 @@
 //! Contains the public `generate_virtual_ts` and `generate_virtual_ts_with_offsets`
 //! functions that orchestrate the full virtual TypeScript generation pipeline.
 
+mod generics;
+mod imports;
+mod options_api;
+mod spans;
+
 use vize_croquis::{BindingType, Croquis, ScopeData, ScopeKind};
 
+use self::generics::{generic_injection_point, references_any_identifier};
+use self::imports::{
+    collect_imported_names, emit_global_component_stubs, emit_reference_type_directives,
+    extract_declared_name,
+};
+use self::options_api::generate_options_api_variables;
+use self::spans::{
+    collect_template_referenced_names, is_local_setup_binding, merge_overlapping_spans,
+    rewrite_export_default_for_module_scope,
+};
 use super::{
     helpers::{
         IMPORT_META_AUGMENTATION, SETUP_SCOPE_HELPER_NAMES, VUE_SETUP_HELPERS, VUE_TYPE_HELPERS,
@@ -20,7 +35,7 @@ use super::{
 use vize_carton::append;
 use vize_carton::cstr;
 use vize_carton::profile;
-use vize_carton::{FxHashMap, FxHashSet, String, camelize, capitalize};
+use vize_carton::{FxHashMap, FxHashSet, String};
 
 /// Generate virtual TypeScript from Vue SFC analysis.
 ///
@@ -938,453 +953,4 @@ pub(crate) fn generate_virtual_ts_with_offsets_and_checks(
     ts.push_str("export default __vize_component__;\n");
 
     VirtualTsOutput { code: ts, mappings }
-}
-
-fn collect_template_referenced_names(summary: &Croquis) -> FxHashSet<String> {
-    let mut names = FxHashSet::default();
-
-    for expression in &summary.template_expressions {
-        collect_expression_identifiers(&mut names, expression.content.as_str());
-        if let Some(guard) = expression.vif_guard.as_ref() {
-            collect_expression_identifiers(&mut names, guard.as_str());
-        }
-    }
-
-    for usage in &summary.component_usages {
-        names.insert(usage.name.as_str().into());
-        if let Some(guard) = usage.vif_guard.as_ref() {
-            collect_expression_identifiers(&mut names, guard.as_str());
-        }
-        for prop in &usage.props {
-            if prop.is_dynamic
-                && let Some(value) = prop.value.as_ref()
-            {
-                collect_expression_identifiers(&mut names, value.as_str());
-            }
-        }
-        for event in &usage.events {
-            if let Some(handler) = event.handler.as_ref() {
-                collect_expression_identifiers(&mut names, handler.as_str());
-            }
-        }
-    }
-
-    for component in &summary.used_components {
-        names.insert(component.as_str().into());
-    }
-
-    for scope in summary.scopes.iter() {
-        match scope.data() {
-            ScopeData::VFor(data) => {
-                collect_expression_identifiers(&mut names, data.source.as_str());
-                if let Some(key_expression) = data.key_expression.as_ref() {
-                    collect_expression_identifiers(&mut names, key_expression.as_str());
-                }
-            }
-            ScopeData::EventHandler(data) => {
-                if let Some(handler) = data.handler_expression.as_ref() {
-                    collect_expression_identifiers(&mut names, handler.as_str());
-                }
-            }
-            _ => {}
-        }
-    }
-
-    names
-}
-
-fn collect_expression_identifiers(names: &mut FxHashSet<String>, expression: &str) {
-    for identifier in vize_croquis::analyzer::extract_identifiers_oxc(expression) {
-        names.insert(identifier.as_str().into());
-    }
-}
-
-fn merge_overlapping_spans(mut spans: Vec<(u32, u32)>) -> Vec<(u32, u32)> {
-    spans.retain(|(start, end)| start < end);
-    spans.sort_by_key(|&(start, end)| (start, end));
-
-    let mut merged: Vec<(u32, u32)> = Vec::with_capacity(spans.len());
-    for (start, end) in spans {
-        if let Some((_, previous_end)) = merged.last_mut()
-            && start <= *previous_end
-        {
-            *previous_end = (*previous_end).max(end);
-            continue;
-        }
-        merged.push((start, end));
-    }
-    merged
-}
-
-fn rewrite_export_default_for_module_scope(text: &str) -> String {
-    let mut output = String::with_capacity(text.len());
-    for segment in text.split_inclusive('\n') {
-        let (line_with_optional_cr, newline) = segment
-            .strip_suffix('\n')
-            .map_or((segment, ""), |line| (line, "\n"));
-        let (line, carriage_return) = line_with_optional_cr
-            .strip_suffix('\r')
-            .map_or((line_with_optional_cr, ""), |line| (line, "\r"));
-
-        let trimmed_line = line.trim_start();
-        if let Some(default_expr) = trimmed_line
-            .strip_prefix("export default")
-            .filter(|rest| rest.chars().next().is_none_or(char::is_whitespace))
-        {
-            let leading_ws = &line[..line.len() - trimmed_line.len()];
-            append!(output, "{leading_ws}const __default__ ={default_expr}");
-        } else {
-            output.push_str(line);
-        }
-        output.push_str(carriage_return);
-        output.push_str(newline);
-    }
-
-    output
-}
-
-fn emit_reference_type_directives(ts: &mut String, script_content: Option<&str>) -> bool {
-    let Some(script) = script_content else {
-        return false;
-    };
-
-    let mut seen = FxHashSet::default();
-    for line in script.lines() {
-        let Some(package) = reference_types_attribute(line) else {
-            continue;
-        };
-        if seen.insert(package) {
-            append!(*ts, "/// <reference types=\"{package}\" />\n");
-        }
-    }
-    !seen.is_empty()
-}
-
-fn reference_types_attribute(line: &str) -> Option<&str> {
-    let line = line.trim_start();
-    if !line.starts_with("///") || !line.contains("<reference") {
-        return None;
-    }
-    attribute_value(line, "types")
-}
-
-fn attribute_value<'a>(line: &'a str, name: &str) -> Option<&'a str> {
-    let needle = cstr!("{name}=");
-    let start = line.find(needle.as_str())? + needle.len();
-    let quote = line[start..].chars().next()?;
-    if quote != '"' && quote != '\'' {
-        return None;
-    }
-    let value_start = start + quote.len_utf8();
-    let value_end = line[value_start..].find(quote)? + value_start;
-    line.get(value_start..value_end)
-}
-
-fn collect_imported_names<'a>(
-    summary: &Croquis,
-    script_content: Option<&'a str>,
-) -> FxHashSet<&'a str> {
-    let Some(script) = script_content else {
-        return FxHashSet::default();
-    };
-
-    summary
-        .import_statements
-        .iter()
-        .flat_map(|imp| {
-            let text = script
-                .get(imp.start as usize..imp.end as usize)
-                .unwrap_or("");
-            extract_import_names(text)
-        })
-        .collect()
-}
-
-fn emit_global_component_stubs(
-    ts: &mut String,
-    summary: &Croquis,
-    options: &VirtualTsOptions,
-    imported_names: &FxHashSet<&str>,
-    enabled: bool,
-) {
-    if !enabled || summary.component_usages.is_empty() {
-        return;
-    }
-
-    let external_template_bindings = options
-        .external_template_bindings
-        .iter()
-        .map(|name| name.as_str())
-        .collect::<FxHashSet<_>>();
-    let auto_import_stub_names = options
-        .auto_import_stubs
-        .iter()
-        .filter_map(|stub| extract_declared_name(stub))
-        .collect::<FxHashSet<_>>();
-
-    let mut emitted_refs = FxHashSet::default();
-    let mut has_header = false;
-    for usage in &summary.component_usages {
-        let name = usage.name.as_str();
-        if summary.bindings.bindings.contains_key(name)
-            || imported_names.contains(&name)
-            || external_template_bindings.contains(&name)
-            || auto_import_stub_names.contains(&name)
-        {
-            continue;
-        }
-
-        let component_ref = to_safe_identifier(name);
-        if !emitted_refs.insert(component_ref.clone()) {
-            continue;
-        }
-
-        if !has_header {
-            ts.push_str("\n// Global component stubs (vue module augmentations)\n");
-            has_header = true;
-        }
-
-        let pascal_name = capitalize(camelize(name).as_str());
-        append!(
-            *ts,
-            "declare const {component_ref}: \"{name}\" extends keyof import(\"vue\").GlobalComponents ? import(\"vue\").GlobalComponents[\"{name}\"]"
-        );
-        if pascal_name.as_str() == name {
-            ts.push_str(" : any;\n");
-        } else {
-            append!(
-                *ts,
-                " : \"{pascal_name}\" extends keyof import(\"vue\").GlobalComponents ? import(\"vue\").GlobalComponents[\"{pascal_name}\"] : any;\n"
-            );
-        }
-    }
-}
-
-/// Extract imported identifier names from an import statement string.
-/// Handles `import { a, b as c } from "..."` and `import D from "..."`.
-/// Returns the local names (e.g., `["a", "c", "D"]`).
-fn extract_declared_name(stub: &str) -> Option<&str> {
-    for prefix in [
-        "declare function ",
-        "declare const ",
-        "declare let ",
-        "declare var ",
-    ] {
-        let Some(rest) = stub.strip_prefix(prefix) else {
-            continue;
-        };
-        let end = rest
-            .find(['<', '(', ':', '=', ';', ' '])
-            .unwrap_or(rest.len());
-        let name = rest[..end].trim();
-        if !name.is_empty() {
-            return Some(name);
-        }
-    }
-
-    None
-}
-
-fn extract_import_names(import_text: &str) -> Vec<&str> {
-    let mut names = Vec::new();
-
-    // Find the content between { and }
-    if let Some(brace_start) = import_text.find('{') {
-        if let Some(brace_end) = import_text.find('}') {
-            let inner = &import_text[brace_start + 1..brace_end];
-            for part in inner.split(',') {
-                let part = part.trim();
-                if part.is_empty() || part.starts_with("//") || part.starts_with("type ") {
-                    continue;
-                }
-                // Handle "name as alias" -> use alias
-                if let Some(as_pos) = part.find(" as ") {
-                    let alias = part[as_pos + 4..].trim();
-                    if !alias.is_empty() {
-                        names.push(alias);
-                    }
-                } else {
-                    // Simple name (strip \r for CRLF files)
-                    let name = part.strip_suffix('\r').unwrap_or(part);
-                    if !name.is_empty() && name.chars().all(|c| c.is_alphanumeric() || c == '_') {
-                        names.push(name);
-                    }
-                }
-            }
-        }
-    } else {
-        // Handle `import Name from "..."`
-        let text = import_text.trim();
-        if let Some(rest) = text.strip_prefix("import ")
-            && let Some(from_pos) = rest.find(" from ")
-        {
-            let name = rest[..from_pos].trim();
-            if !name.is_empty()
-                && !name.contains('{')
-                && !name.contains('*')
-                && name.chars().all(|c| c.is_alphanumeric() || c == '_')
-            {
-                names.push(name);
-            }
-        }
-    }
-
-    names
-}
-
-fn is_local_setup_binding(summary: &Croquis, name: &str) -> bool {
-    let Some(&(start, end)) = summary.binding_spans.get(name) else {
-        return true;
-    };
-
-    !summary
-        .import_statements
-        .iter()
-        .any(|import| start >= import.start && end <= import.end)
-}
-
-// Emit `const <name>: any` declarations for Options API template bindings
-// (`data`/`computed`/`methods`/`inject`/`setup`/`props`, plus any Nuxt 2 globals
-// the legacy path collected). Options API is officially supported in Vue 3, so
-// this is part of the standard build and driven by a runtime opt-in — it costs
-// nothing unless the caller enables Options API / legacy checking.
-fn generate_options_api_variables(
-    mut ts: &mut String,
-    summary: &Croquis,
-    options: &VirtualTsOptions,
-) {
-    let macro_prop_names: FxHashSet<&str> = summary
-        .macros
-        .props()
-        .iter()
-        .map(|prop| prop.name.as_str())
-        .collect();
-    let configured_globals: FxHashSet<&str> = options
-        .template_globals
-        .iter()
-        .map(|global| global.name.as_str())
-        .collect();
-    let mut names: Vec<&str> = summary
-        .bindings
-        .bindings
-        .iter()
-        .filter_map(|(name, binding_type)| {
-            let name = name.as_str();
-            match binding_type {
-                BindingType::Data | BindingType::Options | BindingType::VueGlobal => Some(name),
-                BindingType::Props if !macro_prop_names.contains(name) => Some(name),
-                _ => None,
-            }
-        })
-        .filter(|name| !configured_globals.contains(name))
-        .filter(|name| is_safe_value_identifier(name))
-        .collect();
-    names.sort_unstable();
-    names.dedup();
-
-    if names.is_empty() {
-        return;
-    }
-
-    ts.push_str("  // Options API template bindings\n");
-    for name in &names {
-        append!(ts, "  const {name}: any = undefined as any;\n");
-    }
-    ts.push_str("  ");
-    for name in &names {
-        append!(ts, "void {name};");
-    }
-    ts.push('\n');
-}
-
-fn is_safe_value_identifier(name: &str) -> bool {
-    let mut chars = name.chars();
-    let Some(first) = chars.next() else {
-        return false;
-    };
-    if !(first.is_ascii_alphabetic() || first == '_' || first == '$') {
-        return false;
-    }
-    chars.all(|ch| ch.is_ascii_alphanumeric() || ch == '_' || ch == '$')
-}
-
-fn is_ident_byte(b: u8) -> bool {
-    b == b'_' || b == b'$' || b.is_ascii_alphanumeric()
-}
-
-fn skip_ascii_ws(bytes: &[u8], mut i: usize) -> usize {
-    while i < bytes.len() && bytes[i].is_ascii_whitespace() {
-        i += 1;
-    }
-    i
-}
-
-/// Whether `haystack` mentions any of `idents` as a whole-word identifier.
-/// Used to decide whether a lifted type declaration depends on an SFC generic
-/// parameter and therefore needs that parameter re-declared on it.
-fn references_any_identifier(haystack: &str, idents: &[String]) -> bool {
-    let bytes = haystack.as_bytes();
-    idents.iter().any(|ident| {
-        let ident = ident.as_str();
-        if ident.is_empty() {
-            return false;
-        }
-        let mut from = 0;
-        while let Some(rel) = haystack[from..].find(ident) {
-            let at = from + rel;
-            let before_ok = at == 0 || !is_ident_byte(bytes[at - 1]);
-            let after = at + ident.len();
-            let after_ok = after >= bytes.len() || !is_ident_byte(bytes[after]);
-            if before_ok && after_ok {
-                return true;
-            }
-            from = at + ident.len();
-        }
-        false
-    })
-}
-
-/// Byte offset within a hoisted `type` / `interface` declaration immediately
-/// after the declared name, where synthetic generic parameters can be spliced
-/// in. Returns `None` if the declaration already has its own `<...>` parameter
-/// list or the name can't be located (the declaration is then emitted as-is).
-fn generic_injection_point(decl: &str, type_name: &str) -> Option<usize> {
-    let bytes = decl.as_bytes();
-    let mut i = skip_ascii_ws(bytes, 0);
-
-    // Optional `export` modifier.
-    if decl[i..].starts_with("export")
-        && matches!(bytes.get(i + 6), Some(b) if b.is_ascii_whitespace())
-    {
-        i = skip_ascii_ws(bytes, i + 6);
-    }
-
-    // Declaration keyword.
-    if decl[i..].starts_with("type")
-        && matches!(bytes.get(i + 4), Some(b) if b.is_ascii_whitespace())
-    {
-        i += 4;
-    } else if decl[i..].starts_with("interface")
-        && matches!(bytes.get(i + 9), Some(b) if b.is_ascii_whitespace())
-    {
-        i += 9;
-    } else {
-        return None;
-    }
-    i = skip_ascii_ws(bytes, i);
-
-    // Declared name.
-    if !decl[i..].starts_with(type_name) {
-        return None;
-    }
-    let name_end = i + type_name.len();
-    // Reject partial-name matches (`Foo` inside `Foobar`).
-    if matches!(bytes.get(name_end), Some(&b) if is_ident_byte(b)) {
-        return None;
-    }
-    // Skip declarations that already declare their own type parameters.
-    if bytes.get(skip_ascii_ws(bytes, name_end)) == Some(&b'<') {
-        return None;
-    }
-    Some(name_end)
 }
